Proper exception messages #5

+++ src/Controller/CustomerController.php
@@ -3,11 +3,11 @@
 namespace App\Controller;
 
 use App\Repository\CustomerRepository;
+use ErrorException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
@@ -34,21 +34,27 @@ class CustomerController extends AbstractController
      {
          $data = json_decode($request->getContent(), true);
 
-         $firstName = $data["firstName"];
-         $lastName = $data["lastName"];
-         $email = $data["email"];
-         $phoneNumber = $data["phoneNumber"];
+         try {
 
-         if (empty($firstName) || empty($lastName) || empty($email) || empty($phoneNumber)){
-             throw new NotFoundHttpException("Expecting mandatory parameters:");
+             $firstName = $data["firstName"];
+             $lastName = $data["lastName"];
+             $email = $data["email"];
+             $phoneNumber = $data["phoneNumber"];
 
-         }
+             $id = $this->customerRepository->saveCustomer($firstName, $lastName, $email, $phoneNumber);
 
-         $id = $this->customerRepository->saveCustomer($firstName, $lastName, $email, $phoneNumber);
-         return new JsonResponse(['message' => "Customer created with id number {$id}"], Response::HTTP_CREATED);
+             return new JsonResponse(['message' => "Customer created with id number {$id}"], Response::HTTP_CREATED);
+
+         } catch (\Doctrine\DBAL\Exception\UniqueConstraintViolationException $e) {
+
+             return new JsonResponse(["Error message" => "Enter valid arguments."]);
+
+         } catch (ErrorException $e) {
+
+            return new JsonResponse(["Error message" => "Enter valid arguments"]);
+ 
+        }
      }
-
-
 
     /**
      * @Route("/{id}", name="get_one_customer", methods={"GET"})
@@ -56,6 +62,11 @@ class CustomerController extends AbstractController
     public function getOneCustomer($id): JsonResponse
     {
         $customer = $this->customerRepository->findOneBy(['id' => $id]);
+
+        if (is_null($customer)) {
+
+            return new JsonResponse(["Error message" => "Customer Not found"]);
+        }
 
         $data = [
             'id' => $customer->getId(),
@@ -89,20 +100,24 @@ class CustomerController extends AbstractController
         return new JsonResponse(['customers' => $data], Response::HTTP_OK);
     }
 
-
     /**
      * @Route("/{id}", name="update_customer", methods={"PUT"})
      */
     public function updateCustomer($id, Request $request): JsonResponse
     {
         $customer = $this->customerRepository->findOneBy(['id' => $id]);
+
+        if (is_null($customer)){
+
+            return new JsonResponse(["Error message" => "Customer not found"]);
+        }
+
         $data = json_decode($request->getContent(), true);
 
         $this->customerRepository->updateCustomer($customer, $data);
 
         return new JsonResponse(['message' => 'customer updated with id number '.$id, "data"=>$data]);
     }
-
 
     /**
      * @Route("/{id}", name="delete_customer", methods={"DELETE"})
@@ -111,11 +126,14 @@ class CustomerController extends AbstractController
     {
         $customer = $this->customerRepository->findOneBy(['id' => $id]);
 
+        if (is_null($customer)) {
+
+            return new JsonResponse(["Error message" => "Customer Not Found"]);
+        }
+
         $this->customerRepository->removeCustomer($customer);
 
         return new JsonResponse(['message' => 'customer deleted with id number '.$id]);
     }
-
 }
-
 ?>
